+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 Note: "*" indicates that the commit link is a placeholder. The corresponding link will be updated in the next commit.
 
 
+## v0.0.6
+
+### Features
+- [`62323a0`](https://github.com/wangc9/bike-journey-viewer/commit/62323a047d99651aebcdb50a5e35e501f46ab64b)* Add relations between `Journey` model and `Station` model.
+
+
 ## v0.0.5
 
 ### Features
-- [`9fdd5f6`](https://github.com/wangc9/bike-journey-viewer/commit/9fdd5f6aac53e04adb5c2a0803cd4e305b269d96)* Add `Journey` model.
+- [`62323a0`](https://github.com/wangc9/bike-journey-viewer/commit/62323a047d99651aebcdb50a5e35e501f46ab64b) Add `Journey` model.
 
 
 ## v0.0.4